requirements.txt
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
Pillow==9.2.0
djangorestframework-simplejwt==4.7.2